refactor: :recycle: Move connected validators info response schema to models module
