+++ api/endpoints/validator_models.py
@@ -1,11 +1,8 @@
-from datetime import datetime
 from typing import Any, List, Optional
 from uuid import UUID
 
 from pydantic import BaseModel, Field
 
-from models.agent import Agent
-from models.evaluation import Evaluation
 from models.evaluation_run import EvaluationRunStatus
 from models.openrouter import OpenRouterRuntimeConfig
 from models.problem import ProblemTestResult
@@ -108,15 +105,3 @@ class ValidatorFinishEvaluationRequest(BaseModel):
 
 class ValidatorFinishEvaluationResponse(BaseModel):
     pass
-
-
-class ConnectedValidatorInfo(BaseModel):
-    name: str
-    hotkey: str
-    time_connected: datetime
-
-    time_last_heartbeat: Optional[datetime] = None
-    system_metrics: Optional[SystemMetrics] = None
-
-    evaluation: Optional[Evaluation] = None
-    agent: Optional[Agent] = None
