+++ models/README.md
@@ -1,0 +1,3 @@
+# Models
+
+This module stores all Data Transfer objects used across the project.
